Fixed logic when database is empty

+++ Program.cs
@@ -26,38 +26,45 @@ namespace BrickLinkPoller
             {
                 var databaseOrders = await statusRepository.GetOrders();
 
-                foreach (var order in orders)
+                if (databaseOrders == null)
                 {
-                    if (databaseOrders!.Select(dord => dord.Order_Id).Contains(order.Order_Id))
-                    {
-                        var databaseOrder = databaseOrders!.SingleOrDefault(dord => dord.Order_Id == order.Order_Id);
-                        if (databaseOrder != null)
-                        {
-							if (order.Status == "COMPLETED")
-							{
-								await statusRepository.RemoveOrder(order.Order_Id);
-								message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>) from {order.Store_Name} has been completed!");
-							}
-							else if (databaseOrder.Status != order.Status)
-                            {
-                                var oldStatus = databaseOrder.Status;
-                                databaseOrder.Status = order.Status;
-                                await statusRepository.UpdateOrderStatus(databaseOrder);
-                                message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>) from {order.Store_Name} has changed from {oldStatus} to {databaseOrder.Status}!");
-                                if (order.Tracking_Number != null)
-                                {
-                                    message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>)'s tracking number is {order.Tracking_Number}.");
-                                }
-                            }
-                        }
-					}
-					else
+                    await statusRepository.WriteOrders(orders.Where(o => o.Status != "COMPLETED").ToList());
+                }
+                else
+                {
+					foreach (var order in orders)
 					{
-						await statusRepository.AddOrder(order);
-						message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>) from {order.Store_Name} has been added!");
-						if (order.Tracking_Number != null)
+						if (databaseOrders!.Select(dord => dord.Order_Id).Contains(order.Order_Id))
 						{
-							message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>)'s tracking number is {order.Tracking_Number}.");
+							var databaseOrder = databaseOrders!.SingleOrDefault(dord => dord.Order_Id == order.Order_Id);
+							if (databaseOrder != null)
+							{
+								if (order.Status == "COMPLETED")
+								{
+									await statusRepository.RemoveOrder(order.Order_Id);
+									message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>) from {order.Store_Name} has been completed!");
+								}
+								else if (databaseOrder.Status != order.Status)
+								{
+									var oldStatus = databaseOrder.Status;
+									databaseOrder.Status = order.Status;
+									await statusRepository.UpdateOrderStatus(databaseOrder);
+									message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>) from {order.Store_Name} has changed from {oldStatus} to {databaseOrder.Status}!");
+									if (order.Tracking_Number != null)
+									{
+										message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>)'s tracking number is {order.Tracking_Number}.");
+									}
+								}
+							}
+						}
+						else
+						{
+							await statusRepository.AddOrder(order);
+							message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>) from {order.Store_Name} has been added!");
+							if (order.Tracking_Number != null)
+							{
+								message.AppendLine($"Order [#{order.Order_Id}](<https://www.bricklink.com/orderDetail.asp?ID={order.Order_Id}>)'s tracking number is {order.Tracking_Number}.");
+							}
 						}
 					}
 				}
